# fix: Resolve performance summary data tracking issues (#9)